Test Updates (#17)
* integrate kustomize with spot interrupt test

* fix k8s compatability test

+++ test/spot-termination-test/run-spot-termination-test.sh
@@ -5,11 +5,11 @@ START=$(date +%s)
 SCRIPTPATH="$( cd "$(dirname "$0")" ; pwd -P )"
 PLATFORM=$(uname | tr '[:upper:]' '[:lower:]')
 CLUSTER_NAME_BASE="spot-termination-test"
-CLUSTER_CREATION_TIMEOUT_IN_SEC=90
+CLUSTER_CREATION_TIMEOUT_IN_SEC=300
 TEST_ID=$(uuidgen | cut -d'-' -f1 | tr '[:upper:]' '[:lower:]')
 PRESERVE=false
-TAINT_CHECK_CYCLES=10
-TAINT_CHECK_SLEEP=10
+TAINT_CHECK_CYCLES=15
+TAINT_CHECK_SLEEP=15
 
 DOCKER_ARGS=""
 DEFAULT_NODE_TERMINATION_HANDLER_DOCKER_IMG="node-termination-handler:customtest"
@@ -17,9 +17,14 @@ NODE_TERMINATION_HANDLER_DOCKER_IMG=""
 DEFAULT_EC2_METADATA_DOCKER_IMG="ec2-meta-data-proxy:customtest"
 EC2_METADATA_DOCKER_IMG=""
 
-K8_DEPLOYMENT_SPEC="spot-termination-test.yaml"
+TMP_DIR=$SCRIPTPATH/tmp-$TEST_ID
 
-K8_1_16="kindest/node:v1.16.2@sha256:5bbdfa140633b135672ff0e1eb1a1b37afcab36216103c0b3d97337c62c5e2a1"
+KUSTOMIZATION_FILE="$TMP_DIR/kustomization.yaml"
+NTH_OVERLAY_FILE="nth-image-overlay.yaml"
+METADATA_OVERLAY_FILE="ec2-metadata-image-overlay.yaml"
+REGULAR_POD_OVERLAY_FILE="ec2-metadata-regular-pod-image-overlay.yaml"
+
+K8_1_16="kindest/node:v1.16.2@sha256:2c68d327c2833fa9c0f81b5fd36499cf969646cd50affecd21b4725d37931e21"
 K8_1_15="kindest/node:v1.15.3@sha256:27e388752544890482a86b90d8ac50fcfa63a2e8656a96ec5337b902ec8e5157"
 K8_1_14="kindest/node:v1.14.6@sha256:464a43f5cf6ad442f100b0ca881a3acae37af069d5f96849c1d06ced2870888d"
 K8_1_13="kindest/node:v1.13.10@sha256:2f5f882a6d0527a2284d29042f3a6a07402e1699d792d0d5a9b9a48ef155fa2a"
@@ -83,14 +88,15 @@ done
 CLUSTER_NAME="$CLUSTER_NAME_BASE"-"$TEST_ID"
 echo "🐳 Using Kubernetes $K8_VERSION"
 
-TMP_DIR=$SCRIPTPATH/tmp-$TEST_ID
 ## Append to the end of PATH so that the user can override the executables if they want
 export PATH=$PATH:$TMP_DIR
 mkdir -p $TMP_DIR
 
 function clean_up {
-    [[ "$PRESERVE" == false ]] && kind delete cluster --name $CLUSTER_NAME
-    rm -rf $TMP_DIR
+    if [[ "$PRESERVE" == false ]]; then
+        kind delete cluster --name $CLUSTER_NAME
+        rm -rf $TMP_DIR
+    fi 
 }
 
 function exit_and_fail {
@@ -133,13 +139,35 @@ kind create cluster --name "$CLUSTER_NAME" --image $K8_VERSION --config "$SCRIPT
 export KUBECONFIG="$(kind get kubeconfig-path --name=$CLUSTER_NAME)"
 echo "👍 Created k8s cluster using \"kind\" and added kube config to KUBECONFIG"
 
+cat <<-EOF > $KUSTOMIZATION_FILE
+apiVersion: kustomize.config.k8s.io/v1beta1
+kind: Kustomization
+
+bases:
+- ../../../config/overlays/test
+
+patchesStrategicMerge:
+EOF
+
 if [ -z "$NODE_TERMINATION_HANDLER_DOCKER_IMG" ]; then 
     echo "🥑 Building the node-termination-handler docker image"
     docker build $DOCKER_ARGS -t $DEFAULT_NODE_TERMINATION_HANDLER_DOCKER_IMG --no-cache --force-rm "$SCRIPTPATH/../../." 
     NODE_TERMINATION_HANDLER_DOCKER_IMG="$DEFAULT_NODE_TERMINATION_HANDLER_DOCKER_IMG"
     echo "👍 Built the node-termination-handler docker image"
 else 
-    sed -i "s+$DEFAULT_NODE_TERMINATION_HANDLER_DOCKER_IMG+$NODE_TERMINATION_HANDLER_DOCKER_IMG+" "$SCRIPTPATH/$K8_DEPLOYMENT_SPEC"
+    cat <<-EOF > $TMP_DIR/$NTH_OVERLAY_FILE
+apiVersion: apps/v1
+kind: DaemonSet
+metadata:
+  name: node-termination-handler
+spec:
+  template:
+    spec:
+      containers:
+      - name: node-termination-handler
+        image: $NODE_TERMINATION_HANDLER_DOCKER_IMG
+EOF
+    echo "- $NTH_OVERLAY_FILE" >> $KUSTOMIZATION_FILE
     echo "🥑 Skipping building the node-termination-handler docker image, since one was specified ($NODE_TERMINATION_HANDLER_DOCKER_IMG)"
 fi
 
@@ -149,9 +177,36 @@ if [ -z "$EC2_METADATA_DOCKER_IMG" ]; then
     EC2_METADATA_DOCKER_IMG="$DEFAULT_EC2_METADATA_DOCKER_IMG"
     echo "👍 Built the ec2-meta-data-proxy docker image"
 else 
-    sed -i "s+$DEFAULT_EC2_METADATA_DOCKER_IMG+$EC2_METADATA_DOCKER_IMG+" "$SCRIPTPATH/$K8_DEPLOYMENT_SPEC"
+    cat <<-EOF > $TMP_DIR/$METADATA_OVERLAY_FILE
+apiVersion: apps/v1
+kind: DaemonSet
+metadata:
+  name: node-termination-handler
+spec:
+  template:
+    spec:
+      containers:
+      - name: ec2-metadata-proxy
+        image: $EC2_METADATA_DOCKER_IMG
+EOF
+    echo "- $METADATA_OVERLAY_FILE" >> $KUSTOMIZATION_FILE
+        cat <<-EOF > $TMP_DIR/$REGULAR_POD_OVERLAY_FILE
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: regular-pod-test
+  namespace: default
+spec:
+  template:
+    spec:
+      containers:
+      - name: ec2-meta-data-proxy
+        image: $EC2_METADATA_DOCKER_IMG
+EOF
+    echo "- $REGULAR_POD_OVERLAY_FILE" >> $KUSTOMIZATION_FILE
     echo "🥑 Skipping building the ec2-meta-data-proxy docker image, since one was specified ($EC2_METADATA_DOCKER_IMG)"
 fi
+
 echo "🥑 Tagging worker nodes to execute integ test"
 kubectl label nodes $CLUSTER_NAME-worker lifecycle=Ec2Spot
 kubectl label nodes $CLUSTER_NAME-worker app=spot-termination-test
@@ -163,12 +218,12 @@ kind load docker-image --name $CLUSTER_NAME --nodes=$CLUSTER_NAME-worker,$CLUSTE
 echo "👍 Loaded both images into the cluster"
 
 echo "🥑 Applying the test overlay kustomize config to k8s using kubectl"
-kubectl apply -k "$SCRIPTPATH/../../config/overlays/test"
+kubectl apply -k "$TMP_DIR"
 
 for i in `seq 1 $TAINT_CHECK_CYCLES`; do
-    if kubectl get events | grep regular-pod-test | grep Started; then
+    if kubectl get events | grep Started; then
         echo "✅ Verified regular-pod-test pod was scheduled and started!"
-        if kubectl get nodes $CLUSTER_NAME-worker -o jsonpath="{.spec.taints[].effect}" | grep NoSchedule; then
+        if kubectl get nodes $CLUSTER_NAME-worker | grep SchedulingDisabled; then
             echo "✅ Verified the worker node was cordoned!"
             if [ -z "$(kubectl get pods --namespace=default -o jsonpath="{.items[].spec.nodeName}" | grep worker)" ]; then
                 echo "✅ Verified the regular-pod-test pod was evicted!"
